All entity stuff in org.bukkit moved to org.bukkit.entity
