+++ src/main/java/org/bukkit/entity/Arrow.java
@@ -1,4 +1,4 @@
-package org.bukkit;
+package org.bukkit.entity;
 
 /**
  * Represents an arrow.
