+++ src/main/java/org/bukkit/entity/Egg.java
@@ -1,4 +1,4 @@
-package org.bukkit;
+package org.bukkit.entity;
 
 /**
  * Represents an egg.
