+++ src/main/java/org/bukkit/entity/Snowball.java
@@ -1,4 +1,4 @@
-package org.bukkit;
+package org.bukkit.entity;
 
 /**
  * Implements a snowball.
