+++ src/main/java/org/bukkit/event/block/BlockDamageEvent.java
@@ -2,7 +2,7 @@ package org.bukkit.event.block;
 
 import org.bukkit.Block;
 import org.bukkit.BlockDamageLevel;
-import org.bukkit.Player;
+import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
 
 /**
