+++ src/main/java/org/bukkit/event/block/BlockInteractEvent.java
@@ -1,8 +1,8 @@
 package org.bukkit.event.block;
 
 import org.bukkit.Block;
-import org.bukkit.LivingEntity;
-import org.bukkit.Player;
+import org.bukkit.entity.LivingEntity;
+import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
 
 /**
