+++ src/main/java/org/bukkit/event/block/BlockPlaceEvent.java
@@ -2,7 +2,7 @@ package org.bukkit.event.block;
 
 import org.bukkit.Block;
 import org.bukkit.ItemStack;
-import org.bukkit.Player;
+import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
 
 /**
