+++ src/main/java/org/bukkit/event/entity/EntityCombustEvent.java
@@ -1,6 +1,6 @@
 package org.bukkit.event.entity;
 
-import org.bukkit.Entity;
+import org.bukkit.entity.Entity;
 import org.bukkit.event.Cancellable;
 
 /**
