+++ src/main/java/org/bukkit/event/entity/EntityDamageByBlockEvent.java
@@ -1,7 +1,7 @@
 package org.bukkit.event.entity;
 
 import org.bukkit.Block;
-import org.bukkit.Entity;
+import org.bukkit.entity.Entity;
 import org.bukkit.event.Cancellable;
 import org.bukkit.event.Event;
