+++ src/main/java/org/bukkit/event/player/PlayerItemEvent.java
@@ -4,7 +4,7 @@ import org.bukkit.Block;
 import org.bukkit.BlockFace;
 import org.bukkit.ItemStack;
 import org.bukkit.Material;
-import org.bukkit.Player;
+import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
 
 /**
