+++ src/main/java/org/bukkit/event/player/PlayerMoveEvent.java
@@ -2,7 +2,7 @@
 package org.bukkit.event.player;
 
 import org.bukkit.Location;
-import org.bukkit.Player;
+import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
 import org.bukkit.event.Event;
