+++ src/main/java/org/bukkit/event/vehicle/VehicleCreateEvent.java
@@ -1,6 +1,6 @@
 package org.bukkit.event.vehicle;
 
-import org.bukkit.Vehicle;
+import org.bukkit.entity.Vehicle;
 
 /**
  * Raised when a vehicle is created.
