+++ src/main/java/org/bukkit/event/vehicle/VehicleDamageEvent.java
@@ -1,7 +1,7 @@
 package org.bukkit.event.vehicle;
 
-import org.bukkit.Entity;
-import org.bukkit.Vehicle;
+import org.bukkit.entity.Entity;
+import org.bukkit.entity.Vehicle;
 import org.bukkit.event.Cancellable;
 
 /**
