+++ src/main/java/org/bukkit/event/vehicle/VehicleMoveEvent.java
@@ -1,7 +1,7 @@
 package org.bukkit.event.vehicle;
 
 import org.bukkit.Location;
-import org.bukkit.Vehicle;
+import org.bukkit.entity.Vehicle;
 
 /**
  * Raised when a vehicle moves.
