+++ src/main/java/org/bukkit/fillr/Checker.java
@@ -1,5 +1,6 @@
 package org.bukkit.fillr;
 
+import org.bukkit.entity.Player;
 import java.io.*;
 import java.util.jar.*;
 import org.bukkit.*;
